requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
requests>=2.31.0
playwright>=1.40.0
yt-dlp>=2023.12.0
Pillow>=10.0.0
pydantic>=2.0.0
edge-tts>=6.1.0
moviepy>=1.0.3
tqdm>=4.66.0
